requirements.txt
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
holidays==0.49
idna==3.7
meson==0.63.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
tabulate==0.9.0
urllib3==2.2.1
webencodings==0.5.1